Move DomainWatcher to its own file
The current cache.go contains at least 3
components, let's isolate DomainWatcher for now.
