+++ pkg/virt-handler/cache/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "cache.go",
+        "domain-watcher.go",
         "maps.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/virt-handler/cache",
